+++ changelog.markdown
@@ -8,7 +8,8 @@ Also dates in this file are in [Holocene Calendar] because it is amazing, logica
 ## Unreleased
 ### Fixed
  - Fix the borked recipe that any two items without a recipe together made `Antipolarized Concentrated Pym essence` (#28)
- 
+ - Fix the crash with old Forge (the IRarity class not found error)
+
 ## [1.12.2-2.0.0.1] 12019-12-16
 ### Fixed
  - Fixed startup crashes